MH-10104, Fixed shortcut to build all test modules
The recently introduces -Dtest shortcut to build all test modules may
conflict with Maven tests under some circumstances. This means that it
could for example happen that a build of the matterhorn-json module
fails as it has no tests. Just as a precaution, this patch renames the
shortcut to -Dtests

Signed-off-by: Lars Kiesow <lkiesow@uos.de>