MH-10166, Check for Success State when Probing for Tesseract
The TesseractTextExtractorTest will probe for tesseract by checking that
it return an error code when called with incorrect parameters. This
means that the test will still be executed if something is horribly
wrong with your tesseract.

This commit will fix this behavior by checking for a success code
instead.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>

--

Note: This is a backport of a fix for r/1.5.0