MH-10143, Fix Init Script Error on Debian Based Systems
The problem is that /bin/sh defaults to dash on Debian based systems
which causes errors when executing the start script. Changing the shell
to bash manually works.

This commit just switches the shebang from /bin/sh to /bin/bash which
means that a bash is now required to run the start scripts, but it does
not matter anymore, what the setting for the default shell is.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>