MH-10175, Spaces in Encoding Profile arguments cause errors
Some FFmpeg filters require spaces in the filter configuration. For
example a curves filter definition could look like this:

  -filter:v 'curves=all=0.5/0 0.51/1'

Unfortunately, spaces don't work well with Matterhorn as it tries to
split the filter definition and send them to FFmpeg as a list of
arguments rather than a single argument.

This patch introduce a space-placeholder that can be used to create such
filter graphs. For example, the filter graph defined above would, as
part of an encoding profile, look like this:

  -filter:v curves=all=0.5/0#{space}0.51/1

Signed-off-by: Lars Kiesow <lkiesow@uos.de>

--

Note: This is a backport of a fix for r/1.5.0