MH-10254, Remove GStreamer Composer
This commit removes the GStreamer Composer as #proposed by Rüdiger:

From: Rüdiger Rolf <rrolf@Uni-Osnabrueck.DE>
To: Opencast Matterhorn <matterhorn@opencast.org>
Subject: [Matterhorn] Composer Service relicts #proposal
Date: Fri, 14 Mar 2014 13:56:52 +0100

[…]

1) I would like to propose to remove the current gstreamer composer, as
with all the enhancements that ffmpeg has seen and aware of the problems
that Waldemar had in updating the gstreamer composer last time, where
some of the endpoints for muxing etc do not completely fit to the
structure of gstreamer, it is probably not a direction that we will go
in the future. So I don't feel the need to have an alternative composer
implementation within our basic distribution.

That does NOT mean that services should not use Gstreamer in the future!
I am just talking about the Composer Service. We are about to contribute
the videoeditor which is based in some modules on Gstreamer, and from
our point of view it makes sense there to use Gstreamer. And there may
be other functions where Gstreamer may be superior to ffmpeg.

[…]

Signed-off-by: Lars Kiesow <lkiesow@uos.de>