Remove LTI from default security xml
This patch comments out the OAuth configuration by default. Activating
this should be part of the LTI set-up process and is documented as
such in the LIT guide [l1] which by the by should be part of the release
documentation and not the adopters guide.

This would also mean that adopters not using LTI don't have to
explicitly deactivate it or change the credentials to deny access.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>