MH-10276, Small/fast testing workflows and encoding profiles
The current default workflow contains a lot of operations including some
which are rarely used (e.g. caption upload). This is good as it provides
a basis for creating your own workflow were you just have to remove
unnecessary operations but is bad in the sense that a “normal”
processing takes quite some time.

Thus it would be nice if additional to the default workflow smaller
workflows were available which are usable for fast testing. This patch
adds such workflows accompanied by the necessary encoding profiles.

 - `fast.xml` contains a workflow including optional trimming,
   segmentation and text extraction but no captioning and only a
   conversion to one output format (mp4 with h264 video and aac audio).

 - `minimal.xml` contains the conversion to one output format only.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>