MH-10239, Tracks do not show up in TrimUI
The value for the source flavor of the tracks selected and shown in the
trim-ui was hardcoded as everything containing “work”. This meant that
depending on the workflow tracks would not show up or incorrect ones
would show up.

This patch will go through the operation settings and check which flavors
are actually marked for input and then select the correct tracks
according to this configuration.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>