MH-10283, Forbid trailing spaces and require newline at eof
As #proposed on list (passed 2014-04-29 19:12:41 +0000) all trailing
spaces in the Java code shall be forbidden and every file has to end
with a newline character.

Original Proposal
=================

A while ago we decided to be officially more lax about our code
conventions and allow everything not explicitly forbidden by the
automatick checkstyle rules.  This makes our life easier but brings
also some risks.

Recently I've seen some commits changing the style of a whole file
even though the patch is actually only about a couple of lines in the
code.  So about 90% were just unnecessary changes. Both styles passed
the automatic checkstyle rules and were thus ok, but something like
that messes up the history, makes patches confusing (you have to
search for the actual functionality) and if everything goes wrong, the
original author starts editing the file again and everything is
changed back.

Most of these happened with trailing spaces. We already forbid
trailing spaces in our code and a trailing space will cause a
checkstyle error.  That is, with one exception: Comments.

  To get rid of this problem, I hereby #propose to forbid trailing
  spaces in comments as well.

The second cause of these changes are missing newline characters at
the end of a file. If someone adds content at the end he has to add
this newline character and thus modifying the last line as well
without adding any actual content. Of course these changes will go
into a git patch and show up in the history. Furthermore, most editors
will insert such a newline by default and thus add a modification
automatically.  This is way less important, but still a nuisance.

  To get rid of this problem, I hereby #propose to require newline
  characters at the end of a file.

Both demands should not pose any problem for anybody as all modern IDEs
and editors should be easily configurable to do this automatically.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>