MH-10292 : Prevent series-remote bundle from registering as a service
The service registry analyzes the services present in the system,
and registers then in the database if they define two properties
in their OSGI_INF configuration. This is OK for the bundles that
implement a service directly, but not for the "remote" versions,
which are only proxies to the real services.

The series-remote bundle was incorrectly registering itself in
the service registry, as the provider for a "series.remote"
service. This does not make sense and has been removed.