MH-10276, Reduced maximum GOP size
The default GOP size of 250 causes problems when seeking (using
progressive download) as barely any additional frames are inserted when
using the ultrafast preset which reduces the scene detection of x264 to
its minimum. This patch will force a maximum gop size of 30. This causes
even larger files but as its for testing anyway, it should not matter.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>