MH-10308, Switch default inspection service
With Opencast Matterhorn 1.5 a new ffprobe (the media analysis tool from
the FFmpeg project) based inspection service was introduced to gradually
replace the old mediainfo based implementation. This patch finally
switches the default implementation to the new one by changing some
profiles in the main pom.xml.

The old implementation is still in the alternatives profile and can also
be built from the module directory itself.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>