MH-10313, Incorrect Ingest Parameters cause NullPointerException
Sending invalid data to most of the /ingest/add... endpoints would cause
internal NullPointerExceptions and thus a “500 Internal Server Error” to
be returned to the client.

This patch fixes this behaviour by checking the sanity of the input data
and if necessary and/or sensible returning a “400 Bad Request” instead.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>