MH-10220, Systemd Unit File
Most RedHat based distributions are slowly switching to Systemd in favor
of the classic SysV-Init tool. Fedora uses Systemd already for a couple
of versions, RHEL, CentOS and Scientific Linux will switch to Systemd
with the new version 7.

This commit contains a Systemd unit file utilizing the functionality of
the already included start script. This way Matterhorn can be run using
Systemd directly.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>