MH-10316 Series endpoint shows non-public series when privileges are changed.
The ACLs (SolrFields.ACCESS_CONTROL_CONTRIBUTE, SolrFields.ACCESS_CONTROL_EDIT and SolrFields.ACCESS_CONTROL_READ) are not cleaned up
in the solr input document and it grows indefinitely. So when a series is changed to public it can never be set to private.

Contribution thanks to the support of the U. of Helsinki.