MH-8638 [Improvement] Improve service statistics query
This patch improves the query used to get the service statistics. Instead of a join-query the
query is now executed on only one table (mh_job). Together with an additional index the query is
now executing several times faster than before.

This patch requires an update to the database schema (new index).

Further improvements in this area may be made by caching the service statistics. This is al-
though not part of this patch.