MH-10313, Close InputStreams in IngestService Endpoint
This commit makes sure to properly close input streams at several places
in the ingest service endpoint. In general the streams should now be
closed normally and raise an error if there is a problem. In case of a
previous error, however, IOUtils.closequietly() is used to ensure that
the stream is still properly closed.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>