MH-10339, OpenJDK 8 Unit Test Compatibility
While MH seems to work pretty well with OpenJDK 8 there are some test
failures due to small structural differences in generated JSON and XML.

For example while the output with OpenJDK 7 was ["a":"b","c":"d"] the
output now could be ["c":"d","a":"b"] meaning that the order of the
properties is just a bit different while the actual content is the same.

That should not cause any problems, it will, however, let some tests
fail as they use string comparison to check if some JSON/XML structures
are equal.

This patch will fix those tests to make them check for the actual
content and ignore the order of properties and lists.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>