MH-10374: Adds series id column to mh_search table
This commit adds a column to the mh_search table which will contain the series ID (or NULL) of the episode.
Episodes which predate this functionality will automatically have NULL in this column (per the upgrade script), so the SearchServiceDatabaseImpl class now checks for this upon startup.
This check is relatively inexpensive, and only happens when the system is started.  The column itself is not visible outside of the SearchEntity class, and can be regenerated automatically by NULLing out the column if necessary.