MH-10382, Add UI Element to Unregister Capture Agents
To get rid of old capture agents, you need to use the REST endpoints to
send a DELETE request to the right point. This is a hassle and it would
be nice to have a little button in the capture agent table of the admin
UI to do so … of course only after a confirmation dialog.

This commit adds such an “unregister” button to the table of capture
agents in the admin interface. After clicking it, a confirmation window
shows up and, if the removal is confirmed, the necessary REST API call
is made.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>