MH-10482, Switched Encoding Settings for Composite/Concat
This patch changes the encoding settings used for the concat and
composite workflow operations. In particular, the changes are:

 - Increased video quality
 - Faster video conversion
 - Lossless audio

The changes will cost some space, but even with the recent settings,
files couldn't be published like this. Thus size doesn't matter as the
will be re-encoded for distribution anyway.

One nice side-effect of these changes is that the license problems with
faac shouldn't matter anymore as it's not used anymore and we can get
rid of it. That also means less dependencies.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>