MH-10505, Preview Video for Captioning is Dependent on Videoeditor
The preview video was not generated if the video editor was not
selected. There were also JavaScript code remains of the former
algorithms to determine if a preview needs to be generated left. This
commit will fix that problem by always generating a preview video. It
will also fix MH-9353 as the special property is not necessary anymore.

The downside of this patch is that it might make the default workflow a
bit slower under certain circumstances, but in exchange it is much more
robust which I guess is what we want for this workflow (optimized for
robustness and features, not speed).

Signed-off-by: Lars Kiesow <lkiesow@uos.de>