MH-10515, User Might be Redirected to JSON files after Log-in
Under some circumstances users will be redirected to some JSON data
after login. The reason for this is that MH remembers the first page
that was requested without log-in and redirects the user to that page.
If a user was on an admin page that periodically requests data and the
log-in times out (there are several reasons for that to happen), the
first page that is requested is some data and the user will be
redirected to that data instead of a ui.

This patch will fix this behavior by accepting redirects only to a user
interface (URI contains “.htm”). This simple heuristic should be
sufficient as it should cover nearly all cases and the worst thing that
might happen is that the user will be redirected to the welcome page
which is far better than to end up in some kind of JSON data.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>