MH-10541, Disable Default Matterhorn Adopter Tracking
Currently data about users are sent out to the Matterhorn projects
without asking users if these data should be sent. It can be turned off
in the main configuration file, but once Matterhorn was started, it is
too late to do that.

On Wed, 15 Oct 2014 13:58:14 +0200 a #proposal has passed to deactvate
this option by default since the gathered data are rarely used anyway.

Proposal Thread:
  Subject: Matterhorn proposal Matterhorn Adopter Tracking

Signed-off-by: Lars Kiesow <lkiesow@uos.de>