MH-10508, Ingest Service fails to Handle Episode DC Catalog
An ingest to /ingest/addMediaPackage using the episodeDCCatalog or
episodeDCCatalogUri parameter always failed due to incorrect typecasts
of Catalog to DublinCoreCatalog when trying to set the identifier of the
mediapackage.

As discussed on PR #342, the identifiers of a mediapackage and a
DublinCore catalog should be independent anyway, meaning that we can
safely remove the troublesome code.

Additionally, the documentation has be clarified and some minor issues
have been addressed – or rather have not been stripped from the former
commit.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>