MH-10557, Simplify SQL and Unitize Index Names
This patch tries to simplify the SQL code of the update script by using
`IF EXISTS` statements, extracting index definitions if possible and
unifying the names used for indexes.

For more details about the reasoning have a look at the mailing list
thread at [http://bit.ly/1zie2gz].

In the upgrade script, I've kept the index definition in the table
create statement of mh_job_mh_service_registration as otherwise we would
need a complicated statement to determine if the key already existed on
that table. There is not `IF EXISTS` for indexes.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>