MH-10571, Enable Caching and Exclude Capture
Due to the long time, especially to get the dependencies from Nexus, but
also spent on the capture agent tests, the shippable continuous
integration builds did sometimes time out.

To prevent this, this patch enables caching of the dependencies as well
as it excludes the capture profile from the build process.

Notice: Cherry-picked from develop

Signed-off-by: Lars Kiesow <lkiesow@uos.de>