MH-105777 [bugfix] Fixed in memory bookeeping of hostload when dispatching multiple jobs of the same type at once
When multiple jobs of the same time are dispatched at once, the in-memory host load data strucutres had been updated, but then reset with a fresh copy from the database.

In addition, cleaned up the host load comparator by making sure only working services are being passed in as candidates.