MH-10573, Ignore testSorting of search-service-impl
For some time now the sorting test of the search service implementation
is failing on a regular basis (about ince in 20 builds). That happens
on different machines with different set-ups, different Java versions, …
The issue is tracked as MH-10573.

This is clearly a problem since people (including myself) tend to
ignore things that they are expect to fail and other issues might hence
be overlooked.

From my perspective, the three ways we could go are:
 - Fix the problem
 - Throw out the sorting (broken features are just bugs)
 - Set the test to be ignored

We discussed that in yesterdays technical meeting. While one volunteered
to have a look at this, there was a strong opposition against removing
that feature. That is why I hereby #propose to set the test to be
ignored.

Hopefully, with the removal of Solr (proposal by Tobias) this problem
will go away as well.

A proposal for this was sent out on Wed, 11 Feb 2015 12:16:31 +0100 and will
hopefully pass on Sat, 14 Feb 2015 12:16:31 +0100. This may be merged
afterwards.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>