MH-10188 [Feature] Add Cover Image functionality
The Cover Image service allows you to create images based on a template. It uses
XSLT to produce an SVG image which is finally rendered as a PNG image.

With the CoverImageWorkflowOperationHandler custom cover images for recordings
may be produced. The flexible template solution allows to implement a common
branding on all images.

Used together with the ImageToVideoWorkflowOperationHandler and the
ConcatWorkflowOperationHandler, creating trailers and bumpers for video files is
straight forward.