MH-10804, Add Time Zone to Scheduling UI
When you schedule an event e.g. on the USASK test cluster it is unclear
in which timezone you schedule the recording (local, admin node, ca).
It is actually the local time zone of the browser but it would be nice
to have a small hint about the timezone.

This commit adds a row in which the time zone that will be used is
displayed.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>