MH-10726 [Improvement] Prevent archive from downloading files that already exist in local file system
Instead of loading all files via HTTP the archive uses the workspace to retrieve the file reference.
In most of the cases the file is already part of the working file repository and there is no need to
download the file over HTTP again.

In very few cases where files are not yet part of the working file repository this approach leads to
a small overhead: The file is not only downloaded but also a link or a copy of it needs to be created.