10873, Fix Randomly Failing capture-admin-service-impl tests
The capture-admin-service-impl tests are randomly failing due to
existing database tables which might happen because of new databases
being identified by timestamp only, so that if tests run too fast, two
timestamps might be the same and two separate test might use the same
database file.

This patch adds a large random number to the filename, making it
extremely unlikely that two files are named identical.

This will likely fix the problem but in any case not do any harm.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>