MH-10878, Remove runtime-info-ui uitests
A lot of external JavaScript libraries are included in the uitests
section of the matterhorn-runtime-info-ui module. There have to be
documented or removed.

Taking a closer look, I could not find anything that was still using any
part of this code. Also the last real commit to this section was in
2011. It seems like this can be safely removed.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>