MH-10908 [Workaround] Register servlets in same bundle as HTTP context
In Pax Web < 4.0 an HTTP context cannot be shared across bundles. In order to still use
a common HTTP context, all Servlets and Filters need to be registered in the same
bundle context.

This change may be reverted as soon as we switch to Karaf 4.0 witch comes with Pax Web 4.0.

https://ops4j1.jira.com/browse/PAXWEB-558