MH-10952 : Rename bundle matterhorn-execute-operations to matterhorn-execute-workflowoperation
The bundle matterhorn-execute-operations was developed as a contrib tool before
the workflow operation handlers were taken off the "matterhorn-conductor"
bundle and put in their own indidivual modules. Thus, it does not follow the
same naming schema that it was adopted for the bundles of the other workflow
operation handlers.

This commit changes this bundle's old naming schema.