MH-10852 [Improvement] Upgrade EclipseLink JPA (2.0 → 2.6) and JPA (2.0 → 2.1)
Using JPA inside an OSGi container is meanwhile standardized. Using EclipseLink
as JPA implementation in OSGi is supported by the Eclipse project 'Gemini'.

In accordance with the specification, all classes that need access to an Entity
Manager Factory have been refactored: Gemini JPA directly registers a EMF for
each persitence unit in the OSGi service registry. It's no longer required to
create the EMF in the services.

A major limitation is the fact that all entities of a persitence unit must be
part of the same bundle. Refactoring of some of the entities & bundles was
inevitable.

During the refactoring it became clear that having a centralized 'messages'
service is unhandy. The service has therefore been split up into individual
pieces which now have a clear scope.