MH-10992 : Added explicit charset declaration to "create table" statements
Depending on how the database was created, the tables generated by the script
would not have the correct charset. This commit completes all the "create
table" statements in both the install and update database scripts, so that
they explicitly set the charset to utf-8 and thus avoids problems with improper
character encodings.