MH-11057, Fixed Checkstyle Plugin Requirement
The checkstyle plugin configuration was only used to disable the
plug-in. It was otherwise misconfigured. Most importantly, it also
required a specific version of the plug-in, not depending on the version
in the main pom.xml.

This commit fixes that problem by removing the plug-in requirement and
configuration while disabling the plug-in for the assembly
subdirectories.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>