MH-10908, Auto-Adjust startup.properties
The Karaf Maven Plugin will automatically create a startup.properties
file, adjusted to the specific version. The problem is the
karaf.deployer.spring which conflicts with the Opencast internal Spring
Framework and has to be deactivated.

Instead of hardcoding the startup.properties and shipping it as default
configuration, this patch adds the Maven Antrun Plugin to dynamically
remove the bundle from the generated file.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>