MH-11178, Prevent FFmpeg AAC Encoder Bug
With FFmpeg 2.8, it seems like using `-strict -2 -c:a aac -aac_coder
fast` might lead to FFmpeg hanging during processing. The issue seems to
be fixed in the current git head. Still, since it is in the latest
stable and the option has nearly no advantage, let's remove it to be on
the safe side.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>