MH-11031, Fix File Ingest Through Inbox
The new administrative user interface lacks components to select files
inserted into the inbox as media for a new workflow like the old UI did.

This commit fixes that problem by automatically launch a pre-defined
workflow for these media. By default, the admin-ui default workflow is
used and the media are put into the internal publication status so that
they can be reviewed and further processed.

That way the capabilities of the new ui are used to circumvent the
missing ui components.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>