MH-11196, Fix REST Docs UI
The RESTdocs UI was removed in the new Admin UI and there is no place
where the rest documentations for the services are listed at the moment.
This patch brings back a minimalistic version of the documentation
overview page, integrated into the new admin ui-

This patch is inspired by (and is a dependency of) Pascal's new RESTdoc
design:

  https://bitbucket.org/opencast-community/matterhorn/pull-requests/534

It only provides the basic functionality, not yet the new design as
several dependency problems have to be fixed in advance. The next steps
on this are:

- Create a runtime-dependencies-web bundle to be able to share
  JavaScript libraries among multiple Opencast modules. I will move for
  example angular into this module which relieves us on the admin-ng
  dependency.
- Merge runtime-info-ui and runtime-info-ui-ng since both packages are
  basically doing the same thing and the only real difference is the
  login page.
- Apply new styles to the rest docs. They are dependant on the shared
  JavaScript libraries.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>