MH-11197 Add Crowdin configuration file
The file configures all known locations of translation files in the Opencast
project. It enables to sync the sources as well as the translations with the
CLI tool.

The API credentials need to be stored in a separate file (~/.crowdin.yaml).

For more information about the Crowdin CLI tool, please visit
  https://crowdin.com/page/cli-tool