MH-11185 [Bugfix] Hide assets file size if it is set to value < 1
The file size of the several assets is part of the media package. Very often,
the file size is not set though.

At least for media files, there is the option to use the 'analyze' workflow
operation handler, which sets the file size of analyzed files in the media
package accordingly.