MH-11101 [Bugfix] Fix Static Files Service REST endpoint
It is safe to pass an open InputStream to a JAX-RS response, the framework
is taking care of closing the stream. If the stream is closed within the
endpoint itself, there is no way for JAX-RS to read from that stream when
it comes to sending the data back to the client.