MH-11219, Karaf based Solr configuration
`KARAF_DATA` is used for most of the data modified by Opencast. But this
is not entirely true. There are several points where either the mixing
of local and shared data is problematic for distributed set-ups or where
Opencast writes to files outside of `KARAF_DATA`.

This is one in a set of commits, attempting to fix that problem. It will
centralize the configuration for all Solr indexes and create a default
configuration based on `KARAF_DATA`, making it easier to divide local
data and shared content.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>