MH-11257, Deprecated Mkdocs Config
The pages config in the mkdocs.yml uses the deprecated structure. This
will be removed in the next release of MkDocs. This commit updates the
configuration to the new syntax.

As part of the update, the outdated configuration files and keys has
been removed and the upgrade guide has been merged with the second
upgrade guide which is part of the release notes.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>