MH-8577: Create working alternative to authenticate users against LDAP
The existing LDAP support does not really work if no CAS server is used
alongside the LDAP server.

An alternative implementation has been created, which tries to write as
little new code as possible, using the power of Spring Security instead.

This approach, such as the one taken to configure a CAS server, implies
the usage of an alternative security.xml configuration file, that must
be edited with the specific security information of the LDAP server to
connect to.

This commit has been cherry-picked from the one used for the branch 1.6.x
but includes the necessary adjustments to make it work with the branch
2.1.x