MH-10852 Fix issue with job state inconsistencies
After decoupling the JAX-B and the JPA implementation of the Job
interface, instances of those classes could no longer be casted
(JobJpaImpl <-> JaxbJob). Whenever one needed the JAX-B version or the
JPA version specifically, a new instance which wrapps the original Job
needed to be created.

The original implementation of the service registry (and other central
parts of the system) havilly relied on long-living, mutable Job
references that are passed around and whose state is updated
"automagically". Due to the fact described above, this is no longer true
and lead to strange errors.

With this refactoring JobJpaImpl (no called JpaJob) and JaxbJob are no
longer implementations of Job. There is only one known implementation
called JobImpl. In the context of JPA or JAX-B, a Job instance may be
wrapped in the
corresponding wrapper class.

The Job interface remains mutable since many parts of the system
currently rely on this fact. This is still not ideal as the source of
possible errors is not completelly elliminated and should be fixed in
another refactoring.