MH-11307, Distribution Service Location
The Problem
-----------

The old `dist` profile was moved to the worker distributions with the
introduction of Apache Karaf which now makes the Worker nodes serve the
video files by default. This means that your workers need to be public
and that it might become problematic if you have multiple workers.

The change was introduced since a stand-alone HTTP server could be used
to serve the video files in which case the distribution services would
only need to move the processed files into the correct folder on the NFS
which could possibly give you a performance boost.

The problem here is, that if this is not properly configured, this might
cause a lot of problems. Furthermore, how to configure this properly
seems to be largely undocumented.

This Commit
-----------

This commit changes the Karaf distributions in a way that the
distribution again lives on the presentation (engage) servers which
seems to be the safe option.

People with more knowledge can modify the distributions anyway and still
archive the current configuration.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>