Nexus Location as Maven Var
This commit introduces a Maven variable for the Nexus location. This
makes it possible to easily overwrite this in the local Maven
configuration, effectively pointing to another URL like a local Nexus or
a fixed Node of the distributed Opencast Nexus infrastructure without
the need to ask the GeoIP based Nexus gateway first.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>