MH-9831, Prevent libs from phoning home
Some libraries send out data when first started leading to data about
the system being leaked and even a system failing to start up properly
if no network connection is available. This issue got fixed a while ago
but was re-introduced when switching to Karaf.

This patch addresses the issue by adding configuration keys to prevent
the phone home for some libs as well as deactivating some unused other
libs.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>