MH-10111, Remove Image Conversion Step From TextAnalysisService
The TextAnalyzerServiceImpl.java contains an additional conversion step
to make sure that the incomming image is a TIFF file. This is not
necessary for tesseract. It can handle many different image formats by
itself. It is also not necessary since the TextExtractor will always
extract TIFF images from the videos.

This patch gets rid of the code to decrease the amount of internal
dependencies as well as the code.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>