MH-8577 : Add a new search-based method to authenticate users to the security-sample-ldap file
The previous method offered the possibility to obtain authenticated users
based on their DN. However, they may be cases where the user id (the login
name) is not part of the user DN, and therefore this method will not work
in such cases.

A new "filter based" search has been added, that can be used alongside the
DN-based method, to authenticate users in cases where a plain DN search is
not sufficient or effective.