MH-11372 : Make sure flavor constructor arguments contain no "/" character
The flavor constructor made no sanity check to see if the provided
arguments contained a separator character ("/") which might lead to flavors
like "this/flavor/might/exist".

This commit implements a check in the flavor class constructor and adds a
unit test to demonstrate the constructor fails when certain incorrect
argument combinations are used.