MH-11387 : Fix thumbnails not loading in LTI 'courses' view
The problem was caused by the assumption that the download server URL
path should contain a `/static` component (as the default does). This was
used to determine relative URLs of the thumbnails, by removing the initial
part of the URL and keeping only the substring starting with `static/`.
When the URL contains no `static/` component, unexpected values (only the
last character in the URL) were returned.

Because it is not necessary for the resource URLs to be relative, the
offending lines have simply been deleted.