MH-11391 : Improve flavor parsing and behaviour with "edge cases"
Even though there is no document stating the exact flavor syntax, it is
implicitly assumed that they should always have two non-blank parts
separated with a slash ("/"). The current constructor did not ensure
that the parts are non-empty and did not check that they did not content
any "/" character (thus making it possible to create a flavor with e.g.
`invalid/type` and `validsubtype` that would result in an illegal, but
instantiable `invalid/type/validsubtype` flavor), and therefore required
to be fixed.

On the other hand, the `parseFlavor` method worked in a non-intuitive
way by allowing arguments with more than two slash-separated components
(e.g. "flavor/to/parse") to be successfully parsed, keeping the two
first parts and discarding the rest. It was agreed that `parseFlavor`
should fail for such cases, so this commit addresses that.

Finally, some more unit tests demonstrating the new modifications have
also been included.