MH-11420 : Engage converted "&" characters in URLs to "&amp;"
The engage player incorrectly used jQuery's `html()` function to
retrieve information from the DOM. This caused problems with the stream
security because such function returns the HTML content without
interpreting it, thus expressing the ampersand symbols as `&amp;`
instead of `&` and causing the requests to fail.

This commit fixes this problem by substituting the function `html()`
with the function `text()` in those places where it caused problems.
Because the return value of both functions differ when the element is
empty, the code handling these cases had to be accordingly modified.