MH-11468, Remove Unused Resources
This commit removed several unused resources from the old runtime user
interface, as well as from the shared resources. The runtime-ui was
checked manually, for the shared resources I used:

  find shared-resources -type f | while read f; do
    grep -rq "$(basename "$f")" modules/ shared-resources || \
      git rm "$f";
  done

But I excluded the jQuery image resources for the old admin ui since I
am not sure if the file names for that are generated dynamically. Also,
I expect the old admin ui stuff to be completely removed soon anyway.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>