MH-11518, Language selector is always displayed in system language
The language selectors always use the default locale settings of the
system, Opencast is running on. This would for example mean that an
English user would be presented with all language selections translated
into Chinese, if Chinese is the servers system default language.

This patch will cause each option to be presented in it's own language.
However, this is not perfect since only the main languages seem to be
supported by the Java Locale class and a distinction between
sub-languages is not yet possible.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>