MH-11497: Removed query parameters in workspace URIs
This fixes a bug with the workspace when using stream security, where it
tried to use the whole resource name plus the query string as the
filename. 

Note this is a partial fix: the workspace should be refactored to stop
using `String`s to manipulate URIs. However, this is out of the scope
for the QA period.