MH-11624: Parameter `provider` no longer relevant for User equality
Sometimes the current User reference lacks some information, such as the
user's email, the full name or the "provider". This parameter was,
however, taken into account by the "equals" function, to tell whether or
not two Users were equal. 

This caused problems when a non-privileged user started a workflow: the
user would be the owner, but sometimes, comparisons between the current
user and the workflow owner would fail.

This patch solves this situation by removing the seemingly unimportant
`provider` parameter from the `equals` method, so that only the user
name and the organization are taken into account.